Refactor transformers into modular structure for better readability
