+++ src/features/project/components/ProjectItem.tsx
@@ -1,5 +1,5 @@
 import { Action, ActionPanel, Color, List } from "@raycast/api";
-import type { Project } from "~common/utils/transformers";
+import type { Project } from "~common/transformers/project";
 import { useCurrentSpace } from "~space/hooks/useCurrentSpace";
 import { getProjectImageUrl } from "~common/utils/image";
 import { CommonActionPanel } from "~common/components/CommonActionPanel";
