+++ src/features/pull-request/components/PullRequestDetail.tsx
@@ -2,7 +2,8 @@ import { Image } from "@raycast/api";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import type { Detail, List } from "@raycast/api";
 import type { Entity } from "backlog-js";
-import type { Project, PullRequest, PullRequestComment } from "~common/utils/transformers";
+import type { Project } from "~common/transformers/project";
+import type { PullRequest, PullRequestComment } from "~common/transformers/pull-request";
 import { useCurrentSpace } from "~space/hooks/useCurrentSpace";
 import { getUserIconUrl } from "~common/utils/image";
 import { repositoryOptions } from "~common/utils/queryOptions";
